REFACTOR: Move Voting rules to separate file
