+++ preflibtools/aggregation/__init__.py
@@ -1,0 +1,1 @@
+from preflibtools.aggregation.votingrules import *
